requirements.txt
Django==4.0.6
djangorestframework==3.13.1
requests==2.28.1

# JWT auth
djangorestframework-simplejwt==5.2.0

# PostgreSQL access
psycopg2-binary==2.9.3

# Reversible & Cascading softdelete
django-softdelete==0.9.4

# ENV Configuration
python-decouple==3.6

# Debug and performance check
django-debug-toolbar==3.5.0

# AWS S3
boto3==1.24.29
django-storages==1.12.3

# code linting setup
pre-commit==2.20.0

# Test suit
pytest-django==4.5.2
pytest-factoryboy==2.5.0
Faker==13.15.1

# Code Formatter
black==22.6.0

# Application Server
gunicorn==20.1.0

# Load Testing
locust==2.11.0